[ICU-12809] only .cpp files no .c in ICU library code X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39527 Former-commit-id: 7debeb0dcf8495a33ca1d6772bccfaf7e1607171

+++ icu4c/source/io/ufmt_cmn.h
@@ -26,7 +26,7 @@
 #include "unicode/utf16.h"
 
 #define UFMT_DEFAULT_BUFFER_SIZE 128
-#define MAX_UCHAR_BUFFER_SIZE(buffer) (sizeof(buffer)/(U16_MAX_LENGTH*sizeof(UChar)))
+#define MAX_UCHAR_BUFFER_SIZE(buffer) ((int32_t)(sizeof(buffer)/(U16_MAX_LENGTH*sizeof(UChar))))
 #define MAX_UCHAR_BUFFER_NEEDED(strLen) ((strLen+1)*U16_MAX_LENGTH*sizeof(UChar))
 
 /**
